Move calculate_indicator to indicator.py to avoid circular import
