+++ stock_indicators/indicators/alligator.py
@@ -1,9 +1,8 @@
 from typing import Iterable, Optional, TypeVar
 
 from stock_indicators._cslib import CsIndicator
-from stock_indicators.indicators.common.chain import calculate_indicator
 from stock_indicators.indicators.common.helpers import RemoveWarmupMixin
-from stock_indicators.indicators.common.indicator import Indicator
+from stock_indicators.indicators.common.indicator import Indicator, calculate_indicator
 from stock_indicators.indicators.common.results import IndicatorResults, ResultBase
 from stock_indicators.indicators.common.quote import Quote
